pyproject.toml
[project]
name = "openemr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.83.0",
    "fastapi>=0.132.0",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
]
